[skip] Stopped looking up `User` before passing `Reaction` to an observer (5m)
I'd rather give client code the API of being able to access `user` directly; but it's unnecessary to _resolve_ `user` whenever Houston "hears" a reaction — most of them it will need to ignore.

Also, Houston is raising `PG::ConnectionBad: PQsocket() can't get socket descriptor`, frequently, when resolving `user` to pass it to the observer.

The exception could be resolved by wrapping the sensitive code in `Houston.reconnect { }` — but it would still be better if `Reaction#user` were lazily evaluated.